--- src/main/java/com/Linguine/config/auth/CustomOAuth2UserService.java
@@ -1,62 +1,0 @@
-package com.Linguine.config.auth;
-
-import com.Linguine.domain.member.CustomUserDetails;
-import com.Linguine.domain.member.Member;
-import com.Linguine.repository.MemberRepository;
-import com.Linguine.service.MemberService;
-import lombok.RequiredArgsConstructor;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
-
-import javax.servlet.http.HttpSession;
-import java.util.Collections;
-
-@Service
-@RequiredArgsConstructor
-@Slf4j
-public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequest, OAuth2User> {
-    private final MemberRepository memberRepository;
-
-    private final MemberService memberService;
-    private final HttpSession httpSession;
-
-    @Override
-    public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
-        OAuth2UserService<OAuth2UserRequest, OAuth2User> delegate = new DefaultOAuth2UserService();
-        OAuth2User oAuth2User = delegate.loadUser(userRequest);
-
-        // OAuth2 서비스 id (구글, 카카오, 네이버)
-        String registrationId = userRequest.getClientRegistration().getRegistrationId();
-        // OAuth2 로그인 진행 시 키가 되는 필드 값(PK)
-        String userNameAttributeName = userRequest.getClientRegistration().getProviderDetails().getUserInfoEndpoint().getUserNameAttributeName();
-
-        // OAuth2UserService
-        OAuthAttributes attributes = OAuthAttributes.of(registrationId, userNameAttributeName, oAuth2User.getAttributes());
-        Member member = saveOrUpdate(attributes);
-        httpSession.setAttribute("user", new SessionMember(member)); // SessionUser (직렬화된 dto 클래스 사용)
-//        httpSession.setAttribute("user", new CustomUserDetails(member)); // SessionUser (직렬화된 dto 클래스 사용)
-
-        return new DefaultOAuth2User(Collections.singleton(new SimpleGrantedAuthority(member.getRole().getKey())),
-                attributes.getAttributes(),
-                attributes.getNameAttributeKey())
-                ;
-    }
-
-    // 유저 생성 및 수정 서비스 로직
-    private Member saveOrUpdate(OAuthAttributes attributes){
-        Member user = memberRepository.findByEmail(attributes.getEmail())
-//                .map(entity -> entity.update(attributes.getName(), attributes.getPicture()))
-                .map(entity -> entity.updateOAuth(attributes.getName()))
-                .orElse(attributes.toEntity());
-        return memberRepository.save(user);
-    }
-}
